requirements.txt
cassandra-driver
sendgrid